delete and update protect added

+++ backend/controllers/postController.js
@@ -8,7 +8,8 @@ const User = require('../models/userModel')
 //@route GET/api/posts
 //@access private
 const getPosts = asyncHandler(async (req, res) => {
-  const posts = await Post.find()
+  
+  const posts = await Post.find({user:req.user.id})
   res.status(200).json(posts);
 });
 
@@ -18,12 +19,13 @@ const getPosts = asyncHandler(async (req, res) => {
 //@route POST/api/posts/ceate
 //@access private
 const createPost = asyncHandler(async (req, res) => {
-  if (!req.body.title || !req.body.desc || !req.body.image) {
+  if (!req.body.title || !req.body.description || !req.body.image) {
     res.status(400);
     throw new Error("Please add a text");
   }
 
   const post = await Post.create({
+    user:req.user.id,
     title: req.body.title,
     description : req.body.description,
     image: req.body.image
@@ -47,6 +49,21 @@ const updatePost = asyncHandler(async (req, res) => {
     throw new Error('Post Not Found')
   }
 
+
+  const user = await User.findById(req.user.id)
+
+  //find user availability
+  if(!user){
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  //make sure logged user matches the post user
+  if(post.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
   const updatedPost = await Post.findByIdAndUpdate(req.params.id, req.body, {new:true})
 
   res.status(200).json(updatePost);
@@ -66,6 +83,20 @@ const deletePost = asyncHandler(async (req, res) => {
   if(!post){
     res.status(400)
     throw new Error('Post Not Found')
+  }
+
+  const user = await User.findById(req.user.id)
+
+  //find user availability
+  if(!user){
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  //make sure logged user matches the post user
+  if(post.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await Post.deleteOne()
